Expression: Move Literals into a separate sub-package
